Refactoring. Moved & renamed interfaces.
