+++ app/src/main/java/com/example/schedule/util/EntryEventListener.kt
@@ -1,0 +1,10 @@
+package com.example.schedule.util
+
+import com.example.schedule.data.models.ScheduleInfo
+
+
+interface EntryEventListener {
+    fun onCloseClicked()
+    fun addSchedule(schedule: ScheduleInfo)
+    fun searchLocation()
+}
