+++ app/src/main/java/com/example/schedule/util/LocationListener.kt
@@ -1,0 +1,6 @@
+package com.example.schedule.util
+
+
+interface LocationListener {
+    fun setLocation(location: String)
+}
